Issue #237 renamed package DNAnalzer to analyzer. Gradle ignores classes in packages with capital letters because they are against java naming conventions.
