+++ docs/contributing/Contribution_Guideline.md
@@ -13,7 +13,7 @@
 
 - Clone the Repo by going to your local Git Client in a particular local folder in your local machine by using this
   command with your forked repository link in place of below given link: <br/>
-  `git clone https://github.com/Verisimilitude11/DNAnalyzer`
+  `git clone https://github.com/Verisimilitude11/analyzer`
 - Create a branch using below command.
   `git branch <your branch name>`
 - Checkout to your branch.
