+++ src/main/java/analyzer/BasePairCounter.java
