+++ src/main/java/analyzer/DNATools.java
