+++ src/main/java/analyzer/Main.java
@@ -9,7 +9,7 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package DNAnalyzer;
+package analyzer;
 
 import java.io.IOException;
