+++ src/main/java/analyzer/Properties.java
