+++ src/main/java/analyzer/codon/CodonDataUtils.java
@@ -9,9 +9,9 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package DNAnalyzer.codon;
+package analyzer.codon;
 
-import DNAnalyzer.aminoAcid.*;
+import analyzer.aminoAcid.*;
 
 import java.util.List;
