+++ src/main/java/analyzer/codon/CodonFrame.java
@@ -9,7 +9,7 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package DNAnalyzer.codon;
+package analyzer.codon;
 
 import java.util.Objects;
