+++ src/main/java/analyzer/protein/ProteinAnalysis.java
@@ -9,11 +9,11 @@
  * For further inquiries, please reach out to contact@dnanalyzer.live
  */
 
-package DNAnalyzer.protein;
+package analyzer.protein;
 
 import java.util.List;
 
-import DNAnalyzer.Properties;
+import analyzer.Properties;
 
 /**
  * Prints properties of the proteins in the DNA.
