+++ src/main/java/analyzer/protein/ProteinFinder.java
