+++ src/test/java/analyzer/CodonDataUtilsTest.java
