+++ src/test/java/analyzer/MainTest.java
@@ -1,11 +1,11 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package DNAnalyzer;
+package analyzer;
 
 public class MainTest {
     // @Test
     public void mainClassshouldExist() throws ClassNotFoundException {
-        Class.forName("DNAnalyzer.Main");
+        Class.forName("analyzer.Main");
     }
 }
